fully separated cordination from object and move them into the render elements themselves. add a new 'getBeatScale' function into object to see how long into the beat the time is. slider ball functions correctly, but broken under the current slider gen

+++ osu!trainer/src/components/element/CircleRenderElement.h
@@ -13,8 +13,8 @@ public:
 
 	inline float* GetVertices() { return vertices.data(); }
 
-	inline virtual float GetX(int time) { return x; }
-	inline virtual float GetY(int time) { return y; }
+	inline virtual float GetX(float time) { return x; }
+	inline virtual float GetY(float time) { return y; }
 
 	~CircleRenderElement();
 };
